chore: move Thinking component to Content Parts directory
