+++ client/src/components/Chat/Messages/Content/Parts/Thinking.tsx
